+++ packages/react/src/index.ts
@@ -1,6 +1,6 @@
 export { useChunkUpload } from './useChunkUpload';
 export { useBatchUpload } from './useBatchUpload';
-export { useUploadEcho, useBatchEcho } from './useChunkyEcho';
+export { useUserEcho, useUploadEcho, useBatchEcho } from './useChunkyEcho';
 export { setDefaults, getDefaults, createDefaults } from '@netipar/chunky-core';
 export type { DefaultsScope } from '@netipar/chunky-core';
 export type { ChunkUploadReturn } from './useChunkUpload';